+++ unikmer/cmd/util-io.go
@@ -31,6 +31,9 @@ import (
 	gzip "github.com/klauspost/pgzip"
 )
 
+// BufferSize is size of buffer
+var BufferSize = 65536 // os.Getpagesize()
+
 func outStream(file string, gzipped bool, level int) (*bufio.Writer, io.WriteCloser, *os.File, error) {
 	var w *os.File
 	if file == "-" {
@@ -57,9 +60,9 @@ func outStream(file string, gzipped bool, level int) (*bufio.Writer, io.WriteClo
 		if err != nil {
 			return nil, nil, nil, fmt.Errorf("fail to write %s: %s", file, err)
 		}
-		return bufio.NewWriterSize(gw, os.Getpagesize()), gw, w, nil
+		return bufio.NewWriterSize(gw, BufferSize), gw, w, nil
 	}
-	return bufio.NewWriterSize(w, os.Getpagesize()), nil, w, nil
+	return bufio.NewWriterSize(w, BufferSize), nil, w, nil
 }
 
 func inStream(file string) (*bufio.Reader, *os.File, bool, error) {
@@ -78,7 +81,7 @@ func inStream(file string) (*bufio.Reader, *os.File, bool, error) {
 		}
 	}
 
-	br := bufio.NewReaderSize(r, os.Getpagesize())
+	br := bufio.NewReaderSize(r, BufferSize)
 
 	if gzipped, err = isGzip(br); err != nil {
 		return nil, nil, gzipped, fmt.Errorf("fail to check is file (%s) gzipped: %s", file, err)
@@ -87,7 +90,7 @@ func inStream(file string) (*bufio.Reader, *os.File, bool, error) {
 		if err != nil {
 			return nil, r, gzipped, fmt.Errorf("fail to create gzip reader for %s: %s", file, err)
 		}
-		br = bufio.NewReaderSize(gr, os.Getpagesize())
+		br = bufio.NewReaderSize(gr, BufferSize)
 	}
 	return br, r, gzipped, nil
 }